search-experience-test-crawler/requirements.txt
asyncio
crochet
python-dateutil
psycopg2-binary
redis
requests
scrapy